+++ app/adapters/action.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-osf/adapters/action';