+++ CHANGELOG.md
@@ -38,6 +38,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Moved share button in `file-browser-item` to the `file-browser` toolbar
 - Rename button to have class `primary` instead of `success` on the `file-browser` component
 - What screen sizes columns are displayed in `file-browser` table
+- Rename `action` model to `review-action`
+  - Also rename related adapter, serializer and tests
+  - Added model tests for `review-action`
+  - Customize model adapter
 
 ### Fixed
 - Margins for scrollbar on `file-browser`