+++ CHANGELOG.md
@@ -18,7 +18,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Download
   - Download as zip
   - Upload
+- Original Publication Date is added to preprint.
+  - New `originalPublicationDate` property for `preprint` model.
+  - Add unit tests for `preprint` model.
+  - `validated-input` component is moved from ember-osf-preprints to ember-osf repo and modified.
+    - Change `yarn.lock` to add `ember-bootstrap-datepicker` and `ember-cp-validations` as dependencies.
 - Support button to the HOME navbar
+- Class for small-display on `file-browser`
+- Conditional to check between `files` and `items` in array for file upload between chrome and safari
+- alias in provider model to check if has highlighted subjects
+- `preprintDoiCreated` attribute to the `preprint` model
+- `ember-cli-clipboard` to allow copying to clipboard on more browsers
 
 ### Changed
 - getContents() function for files to use `redirect = true` and `mode = 'render'`
@@ -28,6 +38,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove print margin on ember-ace editor on file-detail page
 - Moved share button in `file-browser-item` to the `file-browser` toolbar
 - Rename button to have class `primary` instead of `success` on the `file-browser` component
+- What screen sizes columns are displayed in `file-browser` table
+- Rename `action` model to `review-action`
+  - Also rename related adapter, serializer and tests
+  - Added model tests for `review-action`
+  - Customize model adapter
+- `meta.total` to `meta.total_pages` in osf-serializer
+
+### Fixed
+- Margins for scrollbar on `file-browser`
+- Clickability on dropzone widget
+- Handle Dropzone enable/disable properly
+
+## [0.12.4] - 2017-12-04
+### Added
+- Final banner images and dates
 
 ## [0.12.3] - 2017-11-29
 ### Added